# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trashure API Documentation",
        "description": "Dokumentasi RESTful API untuk Sistem Pengelolaan Sampah dan Penjemputan Trashure.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Trashure API Base URL"
        }
    ],
    "paths": {
        "/admin/harga-sampah": {
            "get": {
                "tags": [
                    "Admin - Harga Sampah"
                ],
                "summary": "Daftar harga sampah (Admin)",
                "description": "Mengambil daftar harga sampah beserta data jenis sampah terkait. Mendukung pencarian berdasarkan nama jenis sampah, filter berdasarkan status, dan pagination.",
                "operationId": "6e2c8489a6ded7bedd841113bfe54dc8",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian berdasarkan nama jenis sampah",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Plastik"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "aktif",
                            "enum": [
                                "aktif",
                                "nonaktif"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar harga sampah berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar harga & poin sampah berhasil diambil."
                                        },
                                        "data": {
                                            "description": "Data pagination",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Harga Sampah"
                ],
                "summary": "Tambah harga sampah baru (Admin)",
                "description": "Menambahkan data harga dan poin sampah baru. Jika jenis sampah belum ada, akan dibuat otomatis.",
                "operationId": "0f648729cd8cf8c5041ad816d75a04d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_jenis_sampah",
                                    "satuan",
                                    "harga_per_satuan",
                                    "nilai_poin_per_satuan",
                                    "berlaku_mulai",
                                    "status"
                                ],
                                "properties": {
                                    "nama_jenis_sampah": {
                                        "description": "Nama jenis sampah",
                                        "type": "string",
                                        "example": "Plastik",
                                        "maxLength": 100
                                    },
                                    "satuan": {
                                        "description": "Satuan ukur",
                                        "type": "string",
                                        "example": "kg",
                                        "maxLength": 20
                                    },
                                    "harga_per_satuan": {
                                        "description": "Harga per satuan (Rp)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    },
                                    "nilai_poin_per_satuan": {
                                        "description": "Nilai poin per satuan",
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "berlaku_mulai": {
                                        "description": "Tanggal berlaku mulai",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-01"
                                    },
                                    "berlaku_selesai": {
                                        "description": "Tanggal berlaku selesai (opsional)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-12-31",
                                        "nullable": true
                                    },
                                    "status": {
                                        "description": "Status harga sampah",
                                        "type": "string",
                                        "example": "aktif",
                                        "enum": [
                                            "aktif",
                                            "nonaktif"
                                        ]
                                    },
                                    "keterangan": {
                                        "description": "Keterangan tambahan",
                                        "type": "string",
                                        "example": "Harga khusus bulan September",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data harga dan poin sampah berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data harga dan poin sampah berhasil ditambahkan."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/jadwal": {
            "get": {
                "tags": [
                    "Admin - Jadwal"
                ],
                "summary": "Daftar jadwal penjemputan (Admin)",
                "description": "Mengambil semua daftar jadwal penjemputan yang telah dijadwalkan oleh admin.",
                "operationId": "bfe96c1767252cd062f51fbf7590c03d",
                "responses": {
                    "200": {
                        "description": "Daftar jadwal penjemputan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar jadwal penjemputan berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil admin tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profil admin tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/pengajuan": {
            "get": {
                "tags": [
                    "Admin - Pengajuan"
                ],
                "summary": "Daftar pengajuan penjemputan baru (Admin)",
                "description": "Mengambil semua daftar pengajuan penjemputan sampah dari warga dengan status 'diajukan'.",
                "operationId": "5fdafb340f8773572069b2e6bdf5a14d",
                "responses": {
                    "200": {
                        "description": "Daftar pengajuan penjemputan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar pengajuan penjemputan berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profil admin tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profil admin tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/pengajuan/{id}": {
            "get": {
                "tags": [
                    "Admin - Pengajuan"
                ],
                "summary": "Detail pengajuan penjemputan (Admin)",
                "description": "Mengambil rincian pengajuan penjemputan sampah berdasarkan ID pengajuan beserta data warga, detail sampah, dan jadwal jika ada.",
                "operationId": "32bba63ccc34945177a970f0ab218706",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pengajuan Penjemputan",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pengajuan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Detail pengajuan penjemputan berhasil diambil."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pengajuan atau Profil Admin tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pengajuan penjemputan tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout pengguna",
                "description": "Menghapus token akses pengguna yang sedang aktif.",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Logout berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout berhasil."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil pengguna login",
                "description": "Mendapatkan data informasi akun pengguna yang sedang login beserta relasinya.",
                "operationId": "bc2032a379aa97707ccc61511897c120",
                "responses": {
                    "200": {
                        "description": "Data profil user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "warga_budi"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "budi@gmail.com"
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "warga"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "aktif"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Gunakan token Bearer dari endpoint /api/login untuk mengakses endpoint yang dilindungi.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}